requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
chromadb>=0.4.0
numpy>=1.24.0
tiktoken>=0.5.0
unstructured>=0.10.0
PyPDF2>=3.0.0
requests>=2.31.0
pdfminer.six>=20221105
